+++ src/lib.rs
@@ -1,8 +1,9 @@
 mod error;
+mod monitor;
 
 pub mod capturer;
-pub mod duplication_context;
 pub mod model;
 pub mod utils;
 
 pub use error::*;
+pub use monitor::*;